add patch: dmenu_omit - filter omitted entries
dmenu_omit is a filtering tool that appends any supplied entries to a
list of items to filter from the dmenu result. This allows the user to
remove troublesome items that exist in path, that need to exist in path,
but are useless from dmenu.

dmenu_omit is the command line tool to append these items to the list.
dmenu_path was modified to filter based on the items in this list.

dmenu_omitlist is the file stored in your configdir containing these
files.

Makefile was updated to install/uninstall the dmenu_omit